requirements.txt
fastapi>=0.115.14
uvicorn>=0.35.0
python-dotenv>=1.0.1
requests>=2.32.3
pydantic>=2.11.7
yt-dlp>=2025.6.30
openai-whisper>=20250625
google-cloud-storage>=3.2.0
motor>=3.7.1
pymongo>=4.13.2